+++ tomcat/src/main/java/nextstep/jwp/presenstation/LoginHandler.java
@@ -4,8 +4,8 @@ import java.util.Map;
 import java.util.Optional;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.ResponseEntity;
+import org.apache.coyote.http11.http.ContentType;
+import org.apache.coyote.http11.handler.ResponseEntity;
 import org.apache.coyote.http11.handler.AbstractRequestHandler;
 import org.apache.coyote.http11.http.HttpHeaders;
 import org.apache.coyote.http11.http.HttpRequest;
