+++ tomcat/src/main/java/nextstep/jwp/presenstation/RegisterHandler.java
@@ -3,8 +3,8 @@ package nextstep.jwp.presenstation;
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.ResponseEntity;
+import org.apache.coyote.http11.http.ContentType;
+import org.apache.coyote.http11.handler.ResponseEntity;
 import org.apache.coyote.http11.handler.AbstractRequestHandler;
 import org.apache.coyote.http11.http.HttpHeaders;
 import org.apache.coyote.http11.http.HttpMethod;
