+++ tomcat/src/main/java/org/apache/coyote/http11/handler/FileHandler.java
@@ -1,7 +1,6 @@
 package org.apache.coyote.http11.handler;
 
-import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.ResponseEntity;
+import org.apache.coyote.http11.http.ContentType;
 import org.apache.coyote.http11.http.HttpHeaders;
 import org.apache.coyote.http11.http.HttpRequest;
 import org.apache.coyote.http11.http.HttpStatus;
