+++ tomcat/src/main/java/org/apache/coyote/http11/handler/RequestHandler.java
@@ -1,6 +1,6 @@
-package nextstep.jwp.presenstation;
+package org.apache.coyote.http11.handler;
 
-import org.apache.coyote.http11.ResponseEntity;
+import org.apache.coyote.http11.handler.ResponseEntity;
 import org.apache.coyote.http11.http.HttpRequest;
 
 public interface RequestHandler {
